Move out DataSync
