+++ pilots/src/oci-pilot/main.rs
@@ -1,3 +1,4 @@
+mod datasync;
 mod pdm_tests;
 mod podman;
