+++ pilots/src/oci-pilot/podman.rs
@@ -1,63 +1,7 @@
+use crate::datasync::DataSync;
 use flakes::config::itf::InstanceMode;
 use flakes::config::{itf::FlakeConfig, CID_DIR};
 use std::{fs, io::Error, path::PathBuf, process::Command, vec};
-
-/// Data Sync
-///
-struct DataSync {}
-
-impl DataSync {
-    /// Sync static data
-    fn sync_static(&self) -> Result<(), Error> {
-        Ok(())
-    }
-
-    /// Sync layers
-    fn sync_delta(&self) -> Result<(), Error> {
-        Ok(())
-    }
-
-    /// Sync files/dirs specified in target/defaults::HOST_DEPENDENCIES
-    /// from the running host to the target path
-    fn sync_host(&self) -> Result<(), Error> {
-        Ok(())
-    }
-
-    // Prune an image by URI
-    fn prune(&self, uri: &str) -> Result<(), Error> {
-        Ok(())
-    }
-
-    /// Initialise environment
-    fn check_cid_dir(&self) -> Result<PathBuf, Error> {
-        if !CID_DIR.exists() {
-            return Err(Error::new(
-                std::io::ErrorKind::NotFound,
-                format!("CID directory \"{}\" was not found", CID_DIR.as_os_str().to_str().unwrap()),
-            ));
-        }
-
-        if std::fs::metadata(CID_DIR.to_path_buf()).unwrap().permissions().readonly() {
-            return Err(Error::new(
-                std::io::ErrorKind::PermissionDenied,
-                format!("Unable to write to \"{}\" directory", CID_DIR.as_os_str().to_str().unwrap()),
-            ));
-        }
-
-        Ok(CID_DIR.to_owned())
-    }
-
-    /// Flush a cid file
-    fn flush_cid(&self) -> Result<(), Error> {
-        // Probably needs an enum with a different errors
-        Ok(())
-    }
-
-    /// Flush all cids
-    fn flush_all_cids(&self) -> Result<(), Error> {
-        Ok(())
-    }
-}
 
 #[derive(Default)]
 struct DataTracker {
